[Refactor] Strings methods other than MediaType
This refactors all of the remaining utility methods (except those
that convert MediaType and XContentBuilder to String) from the
:server to :libs:opensearch-core library. This commit is to keep the
Strings refactor surface area lean in preparation for cloud native and
serverless refactoring.
