+++ server/src/main/java/org/opensearch/cluster/metadata/MetadataCreateIndexService.java
@@ -257,11 +257,8 @@ public class MetadataCreateIndexService {
      * Validate the name for an index or alias against some static rules.
      */
     public static void validateIndexOrAliasName(String index, BiFunction<String, String, ? extends RuntimeException> exceptionCtor) {
-        if (org.opensearch.common.Strings.validFileName(index) == false) {
-            throw exceptionCtor.apply(
-                index,
-                "must not contain the following characters " + org.opensearch.common.Strings.INVALID_FILENAME_CHARS
-            );
+        if (Strings.validFileName(index) == false) {
+            throw exceptionCtor.apply(index, "must not contain the following characters " + Strings.INVALID_FILENAME_CHARS);
         }
         if (index.isEmpty()) {
             throw exceptionCtor.apply(index, "must not be empty");
