+++ server/src/main/java/org/opensearch/common/Strings.java
@@ -32,7 +32,6 @@
 
 package org.opensearch.common;
 
-import org.apache.lucene.util.BytesRefBuilder;
 import org.opensearch.ExceptionsHelper;
 import org.opensearch.OpenSearchException;
 import org.opensearch.core.common.bytes.BytesReference;
@@ -41,11 +40,6 @@ import org.opensearch.core.xcontent.ToXContent;
 import org.opensearch.core.xcontent.XContentBuilder;
 
 import java.io.IOException;
-import java.util.Arrays;
-import java.util.Set;
-
-import static java.util.Collections.unmodifiableSet;
-import static org.opensearch.common.util.set.Sets.newHashSet;
 
 /**
  * String utility class.
@@ -56,89 +50,7 @@ public class Strings {
 
     public static final String[] EMPTY_ARRAY = org.opensearch.core.common.Strings.EMPTY_ARRAY;
 
-    // ---------------------------------------------------------------------
-    // General convenience methods for working with Strings
-    // ---------------------------------------------------------------------
-
-    /**
-     * Check that the given BytesReference is neither <code>null</code> nor of length 0
-     * Note: Will return <code>true</code> for a BytesReference that purely consists of whitespace.
-     *
-     * @param bytesReference the BytesReference to check (may be <code>null</code>)
-     * @return <code>true</code> if the BytesReference is not null and has length
-     * @see org.opensearch.core.common.Strings#hasLength(CharSequence)
-     */
-    public static boolean hasLength(BytesReference bytesReference) {
-        return (bytesReference != null && bytesReference.length() > 0);
-    }
-
-    /**
-     * Test whether the given string matches the given substring
-     * at the given index.
-     *
-     * @param str       the original string (or StringBuilder)
-     * @param index     the index in the original string to start matching against
-     * @param substring the substring to match at the given index
-     */
-    public static boolean substringMatch(CharSequence str, int index, CharSequence substring) {
-        for (int j = 0; j < substring.length(); j++) {
-            int i = index + j;
-            if (i >= str.length() || str.charAt(i) != substring.charAt(j)) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    // ---------------------------------------------------------------------
-    // Convenience methods for working with formatted Strings
-    // ---------------------------------------------------------------------
-
-    /**
-     * Quote the given String with single quotes.
-     *
-     * @param str the input String (e.g. "myString")
-     * @return the quoted String (e.g. "'myString'"),
-     *         or <code>null</code> if the input was <code>null</code>
-     */
-    public static String quote(String str) {
-        return (str != null ? "'" + str + "'" : null);
-    }
-
-    public static final Set<Character> INVALID_FILENAME_CHARS = unmodifiableSet(
-        newHashSet('\\', '/', '*', '?', '"', '<', '>', '|', ' ', ',')
-    );
-
-    public static boolean validFileName(String fileName) {
-        for (int i = 0; i < fileName.length(); i++) {
-            char c = fileName.charAt(i);
-            if (INVALID_FILENAME_CHARS.contains(c)) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    public static boolean validFileNameExcludingAstrix(String fileName) {
-        for (int i = 0; i < fileName.length(); i++) {
-            char c = fileName.charAt(i);
-            if (c != '*' && INVALID_FILENAME_CHARS.contains(c)) {
-                return false;
-            }
-        }
-        return true;
-    }
-
     private Strings() {}
-
-    public static byte[] toUTF8Bytes(CharSequence charSequence) {
-        return toUTF8Bytes(charSequence, new BytesRefBuilder());
-    }
-
-    public static byte[] toUTF8Bytes(CharSequence charSequence, BytesRefBuilder spare) {
-        spare.copyChars(charSequence);
-        return Arrays.copyOf(spare.bytes(), spare.length());
-    }
 
     /**
      * Return a {@link String} that is the json representation of the provided {@link ToXContent}.
